# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel API Auth",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/wallet/balance": {
            "get": {
                "tags": [
                    "Wallet"
                ],
                "summary": "Get wallet balance",
                "operationId": "2a4a1c7eba45be583fce194ae60a72c1",
                "responses": {
                    "200": {
                        "description": "Wallet balance",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "properties": {
                                                "balance": {
                                                    "type": "number",
                                                    "example": 15000
                                                },
                                                "currency": {
                                                    "type": "string",
                                                    "example": "NGN"
                                                },
                                                "wallet_number": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    },
                    {
                        "apiKey": []
                    }
                ]
            }
        },
        "/api/wallet/deposit": {
            "post": {
                "tags": [
                    "Wallet"
                ],
                "summary": "Initialize deposit",
                "operationId": "506845afd12f94e48ec2a586dfb9ee86",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "amount"
                                ],
                                "properties": {
                                    "amount": {
                                        "description": "Amount in Naira",
                                        "type": "number",
                                        "example": 5000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Deposit initialized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "properties": {
                                                "reference": {
                                                    "type": "string"
                                                },
                                                "authorization_url": {
                                                    "description": "Paystack checkout URL",
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    },
                    {
                        "apiKey": []
                    }
                ]
            }
        },
        "/api/wallet/deposit/{reference}/status": {
            "get": {
                "tags": [
                    "Wallet"
                ],
                "summary": "Check deposit status",
                "operationId": "7a3714a82b33fe3f3c9d2c594c5e9223",
                "parameters": [
                    {
                        "name": "reference",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transaction status"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    },
                    {
                        "apiKey": []
                    }
                ]
            }
        },
        "/api/wallet/transactions": {
            "get": {
                "tags": [
                    "Wallet"
                ],
                "summary": "Get transaction history",
                "operationId": "d7003c8f7d8d61b2b6b5b2d6f2377cc3",
                "responses": {
                    "200": {
                        "description": "List of transactions"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    },
                    {
                        "apiKey": []
                    }
                ]
            }
        },
        "/api/wallet/paystack/webhook": {
            "post": {
                "tags": [
                    "Webhooks"
                ],
                "summary": "Paystack webhook handler",
                "operationId": "a67c80c935e0ffe58d6f06fe50397b4e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "event": {
                                        "type": "string",
                                        "example": "charge.success"
                                    },
                                    "data": {
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Webhook processed"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "apiKey": {
                "type": "apiKey",
                "description": "API Key Authentication",
                "name": "x-api-key",
                "in": "header"
            }
        }
    },
...
}